Selecting verts based on ecc/ang/region is working.
Todo:
Need to be able to make retino_patch objects and populate them.
Currently they use hiResCorverVert.
This still may need to be kept for searching with moving the corners.
But for now i should be able to run the experiments based on the verts selected from the new programs.